pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glueprompt"
version = "0.1.0"
description = "Git-based prompt versioning system using branches for managing prompt templates"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Nicholas Young", email = "nick@nyoung.dev"}
]
keywords = [
    "prompt",
    "versioning",
    "git",
    "templates",
    "jinja2",
    "yaml",
    "prompt-management",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "jinja2>=3.1.6",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "click>=8.3.1",
    "rich>=14.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.urls]
Homepage = "https://github.com/Bioto/glue-prompt"
Documentation = "https://github.com/Bioto/glue-prompt#readme"
Repository = "https://github.com/Bioto/glue-prompt"
Issues = "https://github.com/Bioto/glue-prompt/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["glueprompt*"]
namespaces = false

[project.scripts]
glueprompt = "glueprompt.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]

[tool.ruff]
# Set the target Python version
target-version = "py312"

# Include source files and tests
include = ["glueprompt/**/*.py", "tests/**/*.py", "examples/**/*.py"]

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.egg-info",
    ".pytest_cache",
    ".ruff_cache",
]

# Line length to match Black's default
line-length = 120

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "TID",   # flake8-tidy-imports
    "Q",     # flake8-quotes
    "RET",   # flake8-return
    "PTH",   # flake8-use-pathlib
]

# Ignore specific rules
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function call in argument defaults
    "UP007",  # Use X | Y for type annotations (prefer Optional for now)
    "UP047",  # Use type parameters (prefer TypeVar for compatibility)
]

[tool.ruff.lint.per-file-ignores]
# Allow unused imports in __init__.py files
"__init__.py" = ["F401"]
# Allow print statements in CLI and examples
"glueprompt/cli/**/*.py" = ["T201"]
"examples/**/*.py" = ["T201"]
# Relax rules for test files
"tests/**/*.py" = ["T201", "F841", "E402", "SIM108"]

[tool.ruff.lint.isort]
# Group imports
known-first-party = ["glueprompt"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Respect line endings in the file
line-ending = "auto"
